starknet-client: Update to match new interfaces
