+++ Cargo.toml
@@ -45,7 +45,7 @@ num-bigint = { version = "0.4.3" }
 rand = "0.8"
 
 # === On-Chain === #
-starknet = { git = "https://github.com/renegade-fi/starknet-rs" }
+starknet = "0.5"
 
 # === Networking === #
 libp2p = "0.51"
